# DistSender.Send optimizations.