--- a/kv/dist_sender.go
+++ b/kv/dist_sender.go
@@ -462,12 +462,14 @@ func (ds *DistSender) sendRPC(desc *proto.RangeDescriptor,
 
 // Send implements the client.KVSender interface. It verifies
 // permissions and looks up the appropriate range based on the
-// supplied key and sends the RPC according to the specified
-// options.
-// If the request spans multiple ranges (which is possible for
-// Scan or DeleteRange requests), Send sends requests to the
-// individual ranges sequentially and combines the results
-// transparently.
+// supplied key and sends the RPC according to the specified options.
+//
+// If the request spans multiple ranges (which is possible for Scan or
+// DeleteRange requests), Send sends requests to the individual ranges
+// sequentially and combines the results transparently.
+//
+// This may temporarily adjust the request headers, so the client.Call
+// must not be used concurrently until Send has returned.
 func (ds *DistSender) Send(call client.Call) {
 
 	// TODO: Refactor this method into more manageable pieces.
@@ -481,12 +483,10 @@ func (ds *DistSender) Send(call client.Call) {
 	retryOpts := ds.rpcRetryOptions
 	retryOpts.Tag = "routing " + call.Method().String() + " rpc"
 
-	// responses and descNext are only used when executing across ranges.
-	var responses []proto.Response
 	var descNext *proto.RangeDescriptor
-	// args will be changed to point to a copy of call.Args if the request
-	// spans ranges since in that case we need to alter its contents.
 	args := call.Args
+	reply := call.Reply
+	endKey := args.Header().EndKey
 
 	// In the event that timestamp isn't set and read consistency isn't
 	// required, set the timestamp using the local clock.
@@ -495,7 +495,6 @@ func (ds *DistSender) Send(call client.Call) {
 	}
 
 	for {
-		reply := call.Reply
 		err := util.RetryWithBackoff(retryOpts, func() (util.RetryStatus, error) {
 			reply.Header().Reset()
 			descNext = nil
@@ -503,7 +502,7 @@ func (ds *DistSender) Send(call client.Call) {
 			if err == nil {
 				// If the request accesses keys beyond the end of this range,
 				// get the descriptor of the adjacent range to address next.
-				if desc.EndKey.Less(call.Args.Header().EndKey) {
+				if desc.EndKey.Less(endKey) {
 					if _, ok := call.Reply.(proto.Combinable); !ok {
 						return util.RetryBreak, util.Error("illegal cross-range operation", call)
 					}
@@ -511,7 +510,7 @@ func (ds *DistSender) Send(call client.Call) {
 					// re-run as part of a transaction for consistency. The
 					// case where we don't need to re-run is if the read
 					// consistency is not required.
-					if call.Args.Header().Txn == nil &&
+					if args.Header().Txn == nil &&
 						args.Header().ReadConsistency != proto.INCONSISTENT {
 						return util.RetryBreak, &proto.OpRequiresTxnError{}
 					}
@@ -519,23 +518,12 @@ func (ds *DistSender) Send(call client.Call) {
 					// previous descriptor and range lookups use cache
 					// prefetching.
 					descNext, err = ds.rangeCache.LookupRangeDescriptor(desc.EndKey)
-					// If this is the first step in a multi-range operation,
-					// additionally copy call.Args because we will have to
-					// mutate it as we talk to the involved ranges.
-					if len(responses) == 0 {
-						args = gogoproto.Clone(call.Args).(proto.Request)
-					}
 					// Truncate the request to our current range.
 					args.Header().EndKey = desc.EndKey
 				}
 			}
-			// true if we're dealing with a range-spanning request.
-			isMulti := len(responses) > 0 || descNext != nil
+
 			if err == nil {
-				if isMulti {
-					// Make a new reply object for this call.
-					reply = gogoproto.Clone(call.Reply).(proto.Response)
-				}
 				err = ds.sendRPC(desc, args, reply)
 				if err == nil && reply.Header().Error != nil {
 					err = reply.Header().GoError()
@@ -562,57 +550,55 @@ func (ds *DistSender) Send(call client.Call) {
 						return util.RetryContinue, nil
 					}
 				}
-			} else if isMulti {
-				// If this request spans ranges, collect the replies.
-				responses = append(responses, reply)
+				return util.RetryBreak, err
+			}
 
-				// If this request has a bound, such as MaxResults in ScanRequest,
-				// check whether enough rows are got in this round.
-				if args, ok := args.(proto.Bounded); ok && args.GetBound() > 0 {
-					bound := args.GetBound()
-					if reply, ok := reply.(proto.Countable); ok {
-						if nextBound := bound - reply.Count(); nextBound > 0 {
-							// Update bound for the next round.
-							args.SetBound(nextBound)
-						} else {
-							// Set flag to break the loop.
-							descNext = nil
-						}
+			// If this request has a bound, such as MaxResults in
+			// ScanRequest, check whether enough rows have been retrieved.
+			if args, ok := args.(proto.Bounded); ok && args.GetBound() > 0 {
+				bound := args.GetBound()
+				if reply, ok := reply.(proto.Countable); ok {
+					if nextBound := bound - reply.Count(); nextBound > 0 {
+						// Update bound for the next round.
+						args.SetBound(nextBound)
+					} else {
+						// Set flag to break the loop.
+						descNext = nil
 					}
 				}
+			}
 
-				// descNext can be nil in two cases:
-				// 1. Got enough rows in the middle of the request.
-				// 2. It is the last range of the request.
-				if descNext == nil {
-					// Combine multiple responses into the first one.
-					for _, r := range responses[1:] {
-						// We've already ascertained earlier that we're dealing with a
-						// Combinable response type.
-						responses[0].(proto.Combinable).Combine(r)
-					}
-
-					// Write the final response back.
-					gogoproto.Merge(call.Reply, responses[0])
-				}
+			if call.Reply != reply {
+				// This is a multi-range request. Combine the new response
+				// with the existing response.
+				call.Reply.(proto.Combinable).Combine(reply)
 			}
 			return util.RetryBreak, err
 		})
+
+		// "Untruncate" EndKey to original. We do this even on error in
+		// case the caller will retry using the same args.
+		args.Header().EndKey = endKey
+
 		// Immediately return if querying a range failed non-retryably.
 		// For multi-range requests, we return the failing range's reply.
 		if err != nil {
-			reply.Header().SetGoError(err)
-			gogoproto.Merge(call.Reply, reply) // Only relevant in multi-range case.
+			call.Reply.Header().SetGoError(err)
 			return
 		}
 		// If this was the last range accessed by this call, exit loop.
 		if descNext == nil {
 			break
 		}
+
 		// In next iteration, query next range.
 		args.Header().Key = descNext.StartKey
-		// "Untruncate" EndKey to original.
-		args.Header().EndKey = call.Args.Header().EndKey
+
+		if reply == call.Reply {
+			// This is a mult-range request, make a new reply object for
+			// subsequent iterations of the loop.
+			reply = args.CreateReply()
+		}
 	}
 }