Don't copy the args when processing multi-range requests, but simply
keep track of the original end-key and restore it on each loop
iteration.

Simplify the handling of responses in multi-range requests to that we
combine responses early instead of building up a slice of responses and
combining them at the end.

Only allocate another response when the second range for a multi-range
request is processed instead of allocating a response per range.

benchmark                           old ns/op     new ns/op     delta
BenchmarkClientScan1Version1Row     328910        314646        -4.34%

benchmark                           old MB/s     new MB/s     speedup
BenchmarkClientScan1Version1Row     3.11         3.25         1.05x