@tschottdorf 2015-04-24
LGTM mod the two open comments.
Nice simplification.